backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg[binary]
alembic
pydantic-settings
python-dotenv
httpx
openai
celery
redis
twilio
pytest
pytest-asyncio
actiancortex
